Finished palindrome
Palindrome validation complete
